requirements.txt
streamlit>=1.51.0
requests>=2.32.5
PyPDF2>=3.0.1
python-docx>=1.2.0
python-pptx>=1.0.2
